examples/example-fastmcp-mcp/pyproject.toml
[tool.poetry]
name = "example-fastmcp-mcp"
version = "0.1.0"
description = ""
authors = ["Auth0 <support@auth0.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.1.1"
mcp = "^1.14.1"
auth0-api-python = {path = "../..", develop = true}
starlette = "^0.48.0"
uvicorn = "^0.36.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"